Avoid globals.pre and provide FHS game data dir via macro (requires `SYSINTALL=1`)
Remove the make step to convert globals.pre to a header just for
providing the game data dir by passing that directory as a macro
definition. Additionally make the result visible on the -h output.

To support running the game in place on Linux, and simplifying the
logic, system install with FHS data path should now be opted-in
with `SYSINTALL=1`.

Co-authored-by: Rémi Verschelde <rverschelde@gmail.com>